pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-agent"
version = "0.1.0"
description = "Email inbox agent using OpenAI Agents SDK and Gmail API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.0.0",
  "openai-agents>=0.2.0",
  "google-api-python-client>=2.0.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
]
